Rename to when_policy.hpp to fail_policy.hpp
